characteristics tabular input

+++ backend/controllers/ItemController.php
@@ -2,10 +2,12 @@
 
 namespace backend\controllers;
 
+use common\models\ItemCat;
 use common\models\search\ItemSearch;
+use yii\base\Model;
 use Yii;
 use common\models\Item;
-use backend\models\CharacteristicsForm;
+use common\models\Characteristic;
 use common\models\CharacteristicItem;
 use yii\web\NotFoundHttpException;
 use yii\web\UploadedFile;
@@ -24,14 +26,7 @@ class ItemController extends Controller
     public function actionIndex()
     {
         $searchModel = new ItemSearch();
-        /*$dataProvider = new ActiveDataProvider([
-            'query' => Item::find(),
-            'pagination' => [
-                'pageSize' => 10,
-            ],
-        ]);*/
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
@@ -72,14 +67,11 @@ class ItemController extends Controller
             if($model->save()){
                 if(isset($model->imageFile)) {
                     $filename = $model->imageFile->baseName;
-                    if (file_exists(Item::getPath() . $model->imageFile->baseName . '.' . $model->imageFile->extension)) {
+                    /*if (file_exists(Item::getPath() . $model->imageFile->baseName . '.' . $model->imageFile->extension)) {
                         $filename = $filename . rand(1, 20);
-                    }
+                    }*/
                     $model->imageFile->saveAs(Item::getPath() . $filename . '.' . $model->imageFile->extension);
                 }
-                /*return $this->render('characteristics', [
-                    'model' => $model,
-                ]);*/
                 return $this->redirect(['characteristics', 'id' => $model->id]);
             }
         }
@@ -89,29 +81,38 @@ class ItemController extends Controller
 
     }
 
-    public function actionCharacteristics($id){
-        if (Yii::$app->request->post()) {
-            foreach(Yii::$app->request->post() as $value){
-                if(is_array($value)) {
-                    foreach ($value as $v) {
-                        if(is_array($v)) {
-                            foreach($v as $v2){
-                                /* @var $newCh CharacteristicItem*/
-                                $newCh = new CharacteristicItem();
-                                $newCh->characteristic_id = $v2[0];
-                                $newCh->item_id = $v2[1];
-                                $newCh->value = $v2[2];
-                                $newCh->save();
-                            }
-                        }
-                    }
-                }
+    public function actionUpdatecharacteristics($id)
+    {
+        $characteristics = $this->findModel($id)->getCharacteristicItems()->indexBy('id')->all();
+        if (Model::loadMultiple($characteristics, Yii::$app->request->post()) && Model::validateMultiple($characteristics)) {
+            foreach ($characteristics as $characteristic) {
+                $characteristic->save(false);
             }
             return $this->redirect(['view', 'id' => $id]);
         }
-        return $this->render('characteristics', [
-            'item' => $this->findModel($id), 'model' => new CharacteristicsForm()
-        ]);
+
+        return $this->render('characteristics', ['characteristics' => $characteristics]);
+    }
+
+    public function actionCharacteristics($id){
+        $chars = ItemCat::findOne($this->findModel($id)->category_id)->getCharacteristics()->all();
+        $characteristics = [new CharacteristicItem()];
+        foreach($chars as $char){
+            $char_item = new CharacteristicItem();
+            /* @var $char Characteristic*/
+            $char_item->item_id = $id;
+            $char_item->characteristic_id = $char->id;
+
+            $characteristics[] = $char_item;
+        }
+        if (Model::loadMultiple($characteristics, Yii::$app->request->post()) && Model::validateMultiple($characteristics)) {
+            foreach ($characteristics as $characteristic) {
+                /* @var $characteristic CharacteristicItem*/
+                $characteristic->save();
+            }
+            return $this->redirect(['view', 'id' => $id]);
+        }
+        return $this->render('characteristics', ['characteristics' => $characteristics]);
     }
 
     /**
@@ -125,7 +126,9 @@ class ItemController extends Controller
         $model = $this->findModel($id);
         if ($model->load(Yii::$app->request->post())) {
             if(isset($model->imageFile)){
-                $lastImage = $model->image;
+                if(isset($model->image)) {
+                    $lastImage = $model->image;
+                }
                 $model->imageFile = UploadedFile::getInstance($model, 'imageFile');
                 $model->upload();
             }
@@ -136,10 +139,10 @@ class ItemController extends Controller
                         $filename = $filename . rand(1, 20);
                     }
                     $model->imageFile->saveAs(Item::getPath() . $filename . '.' . $model->imageFile->extension);
-                }
-                if(isset($lastImage)){
-                    if(file_exists(Item::getPath().$lastImage)){
-                        unlink(Item::getPath().$lastImage);
+                    if(isset($lastImage)){
+                        if(file_exists(Item::getPath().$lastImage)){
+                            unlink(Item::getPath().$lastImage);
+                        }
                     }
                 }
             }
@@ -160,8 +163,10 @@ class ItemController extends Controller
     public function actionDelete($id)
     {
         $model = $this->findModel($id);
-        if(file_exists(Item::getPath().$model->image)){
-            unlink(Item::getPath().$model->image);
+        if($model->image != "") {
+            if (file_exists(Item::getPath() . $model->image)) {
+                unlink(Item::getPath() . $model->image);
+            }
         }
         $model->delete();
         return $this->redirect(['index']);
